+++ aquadoggo/src/graphql/types/document_meta.rs
@@ -5,18 +5,22 @@ use async_graphql::Error;
 use dynamic_graphql::{FieldValue, SimpleObject};
 use p2panda_rs::document::traits::AsDocument;
 
-use crate::db::SqlStore;
-use crate::graphql::scalars::{DocumentIdScalar, DocumentViewIdScalar};
-use crate::graphql::utils::{downcast_document_id_arguments, get_document_from_params};
+use crate::graphql::scalars::{DocumentIdScalar, DocumentViewIdScalar, PublicKeyScalar};
+use crate::graphql::utils::downcast_document;
 
-/// The meta fields of a document.
+/// Meta fields of a document, contains id and authorship information.
 #[derive(SimpleObject)]
 pub struct DocumentMeta {
+    /// The document id of this document.
     #[graphql(name = "documentId")]
-    pub document_id: DocumentIdScalar,
+    document_id: DocumentIdScalar,
 
+    /// The document view id of this document.
     #[graphql(name = "viewId")]
-    pub view_id: DocumentViewIdScalar,
+    document_view_id: DocumentViewIdScalar,
+
+    /// The public key of the author who first created this document.
+    owner: PublicKeyScalar,
 }
 
 impl DocumentMeta {
@@ -24,28 +28,23 @@ impl DocumentMeta {
     ///
     /// Requires a `ResolverContext` to be passed into the method.
     pub async fn resolve(ctx: ResolverContext<'_>) -> Result<Option<FieldValue<'_>>, Error> {
-        let store = ctx.data_unchecked::<SqlStore>();
+        // Parse the bubble up parent value.
+        let document = downcast_document(&ctx);
 
-        // Downcast the parameters passed up from the parent query field
-        let (document_id, document_view_id) = downcast_document_id_arguments(&ctx);
-
-        // Get the whole document
-        let document = get_document_from_params(store, &document_id, &document_view_id).await?;
-
-        // Construct `DocumentMeta` and return it. We defined the document meta
-        // type and already registered it in the schema. It's derived resolvers
-        // will handle field selection.
-        let field_value = match document {
-            Some(document) => {
-                let document_meta = Self {
-                    document_id: document.id().into(),
-                    view_id: document.view_id().into(),
-                };
-                Some(FieldValue::owned_any(document_meta))
-            }
-            None => Some(FieldValue::NULL),
+        // Extract the document in the case of a single or paginated request.
+        let document = match document {
+            super::DocumentValue::Single(document) => document,
+            super::DocumentValue::Paginated(_, _, document) => document,
         };
 
-        Ok(field_value)
+        // Construct `DocumentMeta` and return it. We defined the document meta
+        // type and already registered it in the schema.
+        let document_meta = Self {
+            document_id: document.id().into(),
+            document_view_id: document.view_id().into(),
+            owner: document.author().to_owned().into(),
+        };
+
+        Ok(Some(FieldValue::owned_any(document_meta)))
     }
 }